third_party/WebKit/LayoutTests/inspector/sources/debugger/show-function-definition-expected.txt

Tests that "Show Function Definition" jumps to the correct location.


Running: testRevealFunctionDefinition
Function location revealed: [7:18]

Running: testDumpFunctionDefinition
jumpToMe
jumpToMe()
